Implemented game stats collecting and printing. Fixes.

+++ MTSim.Objects/Animals/Herbivores/Herbivore.cs
@@ -5,7 +5,7 @@ namespace MTSim.Objects.Animals.Herbivores
 {
     public abstract class Herbivore : Animal
     {
-        protected Herbivore(int id, Island island, Point coords, int maxSpeed, double maxSatiety, double weight, Dictionary<string, double> food)
+        protected Herbivore(long id, Island island, Point coords, int maxSpeed, double maxSatiety, double weight, Dictionary<string, double> food)
             : base(id, island, coords, maxSpeed, maxSatiety, weight, food)
         {
         }
